requirements.txt
sty
colorcet
numpy
matplotlib
qiskit
qiskit-aer

git+https://github.com/learnqc/code.git

panel
openai
python-dotenv
